Moved to use shared lib
