+++ src/vips-notification-worker/pom.xml
@@ -19,6 +19,14 @@
 		<java.version>1.8</java.version>
 	</properties>
 
+
+	<repositories>
+		<repository>
+			<id>dps-group</id>
+			<url>http://nexus-pjztlm-tools.pathfinder.gov.bc.ca/repository/dps-group/</url>
+		</repository>
+	</repositories>
+
 	<dependencies>
 
 		<dependency>
@@ -77,9 +85,9 @@
 		</dependency>
 
 		<dependency>
-			<groupId>ca.bc.gov.open.pssg.rsbc</groupId>
-			<artifactId>vips-ords-client</artifactId>
-			<version>0.4.0-SNAPSHOT</version>
+			<groupId>ca.bc.gov.open</groupId>
+			<artifactId>jag-vips-client</artifactId>
+			<version>0.0.2-SNAPSHOT</version>
 		</dependency>
 
 		<dependency>
